database schema respects the default currency
when default currency is changed. the amount column in the database
has a postfix of the subunit of the currency so set

likewise for the currency column as well. It also changes the
default value to the currencies iso code

add gem debugger for ruby 1.9 on MRI ruby

fix two failing tests where the default_currency is set to EUR
in spec/dummy/config/initializers/money.rb