Support multiple attributes w/ one call for AR
Adjust usage of `extract_options!` to accepts multiple attributes and
provide a more Rails-y behavior (similar to `validates`, `delegate`,
etc.) It's just moving all the original code inside a loop.

Ex:

```ruby
class Product
  monetize :field_one, :field_two, :field_three, :option => 'value'
end
```

This will raise an error if multiple attributes are used with the `:as`
option. That's intentional since it's an attempt to name multiple
attributes with the same name. All other options should be ok.