Revert "make humanized_money helper respect currency decimal mark and thousands separator"
The formatting of the number depends on the language in the real world,
not on the currency. The best example is the Euro, where it previously
enforce the German format on the entire Euro zone. France, Italy and
Spain for example write `1 234,56 €`, not `1.234.56 €`. And I think,
English speaking countries would expect `1,234.56 €`, even though there
is no English speaking country who uses the Euro. In return, we Germans
also write `$ 1.234,56`, not `$ 1,234.56`. As I see it, you can't show
currencies on a site in German and French now without always passing
the params anymore.

This reverts commit f2dbc6031a25780e29c9d74524c3233394b3c652.